+++ src/main/java/com/aipiabackend/product/service/dto/ProductCreateCommand.java
@@ -1,10 +1,8 @@
 package com.aipiabackend.product.service.dto;
 
-import java.math.BigDecimal;
-
 public record ProductCreateCommand(
     String name,
-    BigDecimal price,
+    Long price,
     Integer stock,
     String description
 ) {
